refactor: split data/storage.py into _storage_* modules
Extract history, template, and style sample CRUD into separate
_storage_*.py files under data/. storage.py now re-exports all
public symbols for backward compatibility.

- _storage_base.py: get_db(), close_db(), schema, db singleton
- _storage_history.py: history CRUD
- _storage_templates.py: template CRUD
- _storage_samples.py: style sample CRUD + FTS5 search
